NEW: Internal transfer (Moved Double to Decimal) (MOVEU-15942)

+++ Project/SantanderTests/Modifiers/TransferOperatives/IternalTransfer/Representables/ExchangeRateRepresentableMock.swift
@@ -11,11 +11,11 @@ import SANPLLibrary
 struct ExchangeRateRepresentableMock: ExchangeRateRepresentable {
     var currency: String
     var currencySymbol: String
-    var buyRate: Double
-    var sellRate: Double
+    var buyRate: Decimal
+    var sellRate: Decimal
     var decPlaces: Int
     
-    public init(currency: String, currencySymbol: String, buyRate: Double, sellRate: Double, decPlaces: Int) {
+    public init(currency: String, currencySymbol: String, buyRate: Decimal, sellRate: Decimal, decPlaces: Int) {
         self.currency = currency
         self.currencySymbol = currencySymbol
         self.buyRate = buyRate
